chore: move cluster config interfaces to own files
