+++ src/commands/cluster/config_interfaces/cluster_ref_connect_config_class.ts
@@ -1,0 +1,9 @@
+/**
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export interface ClusterRefConnectConfigClass {
+  cacheDir: string;
+  devMode: boolean;
+  context: string;
+  clusterRef: string;
+}
