+++ src/commands/cluster/config_interfaces/cluster_ref_default_config_class.ts
@@ -1,0 +1,8 @@
+/**
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export interface ClusterRefDefaultConfigClass {
+  cacheDir: string;
+  devMode: boolean;
+  clusterRef: string;
+}
