+++ src/commands/cluster/config_interfaces/cluster_reset_config_class.ts
@@ -1,0 +1,7 @@
+/**
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export interface ClusterResetConfigClass {
+  clusterName: string;
+  clusterSetupNamespace: string;
+}
